pyproject.toml
[project]
name = "jax-to-tfjs"
version = "0.1.0"
description = "JAX/Flax 모델을 TensorFlow.js로 변환하는 데모"
authors = []
dependencies = [
    "jax>=0.4.20",
    "flax>=0.8.1",
    "tensorflow>=2.15.0",
    "tensorflowjs>=4.15.0",
    "orbax-checkpoint>=0.4.1",
    "tensorflow-datasets>=4.9.4",
    "optax>=0.1.7",
    "inquirer>=3.1.3",
    "tqdm>=4.66.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "pytest>=8.3.5",
    "matplotlib>=3.8.0",
    "scikit-learn>=1.6.1",
    "pyinquirer>=1.0.3",
    "ruff>=0.11.5",
    "seaborn>=0.13.2",
    "pandas>=2.2.3",
]

[project.scripts]
train-jax = "jax_to_tfjs.models.jax_mnist_cnn:train_and_evaluate"
train-flax = "jax_to_tfjs.models.flax_mnist_cnn:train_and_evaluate"
evaluate = "evaluate:main"
convert-to-tfjs = "jax_to_tfjs.convert_to_tfjs:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.1",
    "typer-cli>=0.0.13",
    "pyinquirer>=1.0.3",
]

